football-score-combination/requirements.txt
fastapi
strawberry-graphql
uvicorn
pytest